+++ src/main/groovy/org/jenkinsci/gradle/plugins/jpi/restricted/RestrictedApiException.groovy
@@ -1,0 +1,7 @@
+package org.jenkinsci.gradle.plugins.jpi.restricted
+
+class RestrictedApiException extends RuntimeException {
+    RestrictedApiException() {
+        super('Restricted APIs were detected - see https://tiny.cc/jenkins-restricted')
+    }
+}